requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
twilio>=8.0.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
